fet: get-permissions simpled

+++ routers/files.py
@@ -24,8 +24,10 @@ async def create_file(authorize: auth_middlewares.AuthJWT = Depends(),
                       file: UploadFile = File(...)):
     authorize.jwt_required()
     current_user = await auth_middlewares.get_user(authorize.get_jwt_subject(), _id_check=True)
-    if current_user.company_id is None:
-        raise HTTPException(status_code=400, detail="No company is attached")
+    if current_user.company_id is None or not \
+            (await auth_middlewares.get_permissions(current_user.role_id))['can_upload_files']:
+        raise HTTPException(status_code=400, detail='Company is not attached to the user'
+                                                    ' or does not have permissions for that action')
     file_read = await file.read()
     with open(f'files/{file.filename}', 'wb') as doc:
         doc.write(file_read)
@@ -55,8 +57,10 @@ async def create_file(authorize: auth_middlewares.AuthJWT = Depends(),
 async def add_file_info(data: ItemAddFileInfo, authorize: auth_middlewares.AuthJWT = Depends()):
     authorize.jwt_required()
     current_user = await auth_middlewares.get_user(authorize.get_jwt_subject(), _id_check=True)
-    if current_user.company_id is None:
-        raise HTTPException(status_code=400, detail="No company is attached")
+    if current_user.company_id is None or not \
+            (await auth_middlewares.get_permissions(current_user.role_id))['can_upload_files']:
+        raise HTTPException(status_code=400, detail='Company is not attached to the user'
+                                                    ' or does not have permissions for that action')
     item_updated = dict()
     file_id = data.file_id
 
@@ -115,8 +119,10 @@ async def delete_file(file_id, authorize: auth_middlewares.AuthJWT = Depends()):
         raise HTTPException(status_code=400, detail='Not valid Object_id')
     file = config.db.files.find_one({"_id": ObjectId(file_id)})
     current_user = await auth_middlewares.get_user(authorize.get_jwt_subject(), _id_check=True)
-    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id:
-        raise HTTPException(status_code=400, detail="No company is attached")
+    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id or not \
+            (await auth_middlewares.get_permissions(current_user.role_id))['can_upload_files']:
+        raise HTTPException(status_code=400, detail='Company is not attached to the user'
+                                                    ' or does not have permissions for that action')
     config.db.files.remove({"_id": ObjectId(file_id)})
     try:
         os.remove(file['path'])
@@ -133,8 +139,10 @@ async def get_file(file_id, authorize: auth_middlewares.AuthJWT = Depends()):
         raise HTTPException(status_code=400, detail='Not valid Object_id')
     file = config.db.files.find_one({"_id": ObjectId(file_id)})
     current_user = await auth_middlewares.get_user(authorize.get_jwt_subject(), _id_check=True)
-    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id:
-        raise HTTPException(status_code=400, detail="No company is attached")
+    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id or not \
+            (await auth_middlewares.get_permissions(current_user.role_id))['can_download_files']:
+        raise HTTPException(status_code=400, detail='Company is not attached to the user'
+                                                    ' or does not have permissions for that action')
     file = await files_additional_funcs.delete_object_ids_from_dict(file)
     return file
 
@@ -148,8 +156,11 @@ async def get_file_by_third_party_id(third_party_id, authorize: auth_middlewares
                                                   {"$match": {
                                                       "third_parties.third_party_id": ObjectId(third_party_id)}}]))
     current_user = await auth_middlewares.get_user(authorize.get_jwt_subject(), _id_check=True)
-    if current_user.company_id is None or len(company) != 1 or str(company[0]['_id']) != current_user.company_id:
-        raise HTTPException(status_code=400, detail="No company is attached")
+    if current_user.company_id is None or len(company) != 1 \
+            or str(company[0]['_id']) != current_user.company_id or not \
+            (await auth_middlewares.get_permissions(current_user.role_id))['can_download_files']:
+        raise HTTPException(status_code=400, detail='Company is not attached to the user'
+                                                    ' or does not have permissions for that action')
     company = company[0]
 
     files_not_in_folders = config.db.files.find({
@@ -219,8 +230,11 @@ async def get_file_by_third_party_folder_id(third_party_folder_id, authorize: au
          {"$unwind": "$third_parties.folders"},
          {"$match": {"third_parties.folders.third_party_folder_id": ObjectId(third_party_folder_id)}}]))
     current_user = await auth_middlewares.get_user(authorize.get_jwt_subject(), _id_check=True)
-    if current_user.company_id is None or len(company) != 1 or str(company[0]['_id']) != current_user.company_id:
-        raise HTTPException(status_code=400, detail="No company is attached")
+    if current_user.company_id is None or len(company) != 1 \
+            or str(company[0]['_id']) != current_user.company_id or not \
+            (await auth_middlewares.get_permissions(current_user.role_id))['can_download_files']:
+        raise HTTPException(status_code=400, detail='Company is not attached to the user'
+                                                    ' or does not have permissions for that action')
 
     files_in_folder = config.db.files.find({
             'third_party_folder_id': ObjectId(third_party_folder_id)})
@@ -257,8 +271,10 @@ async def get_file_history(file_id, authorize: auth_middlewares.AuthJWT = Depend
         raise HTTPException(status_code=400, detail='Not valid Object_id')
     file = config.db.files.find_one({"_id": ObjectId(file_id)})
     current_user = await auth_middlewares.get_user(authorize.get_jwt_subject(), _id_check=True)
-    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id:
-        raise HTTPException(status_code=400, detail="No company is attached")
+    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id or not \
+            (await auth_middlewares.get_permissions(current_user.role_id))['can_download_files']:
+        raise HTTPException(status_code=400, detail='Company is not attached to the user'
+                                                    ' or does not have permissions for that action')
     history_list = list()
     history_list.append(file)
     if file['parent_id'] is not None:
@@ -274,8 +290,10 @@ async def get_uploaded_preview_file(file_id, authorize: auth_middlewares.AuthJWT
     authorize.jwt_required()
     file = config.db.files.find_one({"_id": ObjectId(file_id)})
     current_user = await auth_middlewares.get_user(authorize.get_jwt_subject(), _id_check=True)
-    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id:
-        raise HTTPException(status_code=400, detail="No company is attached")
+    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id or not \
+            (await auth_middlewares.get_permissions(current_user.role_id))['can_download_files']:
+        raise HTTPException(status_code=400, detail='Company is not attached to the user'
+                                                    ' or does not have permissions for that action')
     try:
         return StreamingResponse(open(file['preview_path'], mode="rb"))
 
@@ -289,8 +307,10 @@ async def get_uploaded_file(file_id, authorize: auth_middlewares.AuthJWT = Depen
     authorize.jwt_required()
     file = config.db.files.find_one({"_id": ObjectId(file_id)})
     current_user = await auth_middlewares.get_user(authorize.get_jwt_subject(), _id_check=True)
-    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id:
-        raise HTTPException(status_code=400, detail="No company is attached")
+    if current_user.company_id is None or file is None or str(file['company_id']) != current_user.company_id or not \
+            (await auth_middlewares.get_permissions(current_user.role_id))['can_download_files']:
+        raise HTTPException(status_code=400, detail='Company is not attached to the user'
+                                                    ' or does not have permissions for that action')
     try:
         return FileResponse(file['path'],
                             media_type=file['content_type'],
